Background task for updating draft from xml
